webserver/requirements.txt
Pillow
redis
python-multipart
numpy